+++ build.rs
@@ -2,7 +2,10 @@ use vergen::*;
 
 fn main() {
     let flags = ConstantsFlags::all();
-    gen(flags).expect("Unable to generate the cargo keys!");
+    gen(flags).unwrap_or_else(|_e| {
+        println!("cargo:rustc-env=VERGEN_GIT_BRANCH=unknown");
+        println!("cargo:rustc-env=VERGEN_GIT_SHA=unknown");
+    });
 
     let mut features = vec![];
     for (k, _v) in std::env::vars() {
@@ -11,8 +14,8 @@ fn main() {
             "CARGO_FEATURE_HDF5_OUTPUT" => features.push("hdf5-output"),
             "CARGO_FEATURE_WITH_MPI" => features.push("with-mpi"),
             "CARGO_FEATURE_COMPENSATING_CHIRP" => features.push("compensating-chirp"),
-            "CARGO_FEATURE_NO_RADIATION_REACTION" => features.push("no-radiation-reaction"),
             "CARGO_FEATURE_COS2_ENVELOPE_IN_3D" => features.push("cos2-envelope-in-3d"),
+            "CARGO_FEATURE_LEADING_ORDER_ONLY" => features.push("leading-order-only"),
             _ => {}
         }
     }