+++ src/field/fast_plane_wave.rs
@@ -103,9 +103,15 @@ impl Field for FastPlaneWave {
     }
 
     #[allow(non_snake_case)]
-    fn radiate<R: Rng>(&self, r: FourVector, u: FourVector, dt: f64, rng: &mut R) -> Option<FourVector> {
+    fn radiate<R: Rng>(&self, r: FourVector, u: FourVector, dt: f64, rng: &mut R) -> Option<(FourVector, FourVector)> {
         let (E, B) = self.fields(r);
-        FastFocusedLaser::emit_photon(u, E, B, dt, rng)
+        FastFocusedLaser::emit_photon(u, E, B, dt, rng).map(|k| (k, u - k))
+    }
+
+    #[allow(non_snake_case)]
+    fn pair_create<R: Rng>(&self, r: FourVector, ell: FourVector, dt: f64, rng: &mut R, rate_increase: f64) -> (f64, Option<(FourVector, FourVector)>) {
+        let (E, B) = self.fields(r);
+        FastFocusedLaser::create_pair(ell, E, B, dt, rng, rate_increase)
     }
 }