+++ src/field/focused_laser.rs
@@ -1,10 +1,12 @@
 use std::f64::consts;
 use rand::prelude::*;
+use rand_distr::StandardNormal;
 
 use crate::field::{Field, Polarization};
 use crate::constants::*;
 use crate::geometry::FourVector;
 use crate::nonlinear_compton;
+use crate::pair_creation;
 
 /// Represents the envelope of a focusing laser pulse, i.e.
 /// the field after cycle averaging
@@ -14,6 +16,7 @@ pub struct FocusedLaser {
     duration: f64,
     wavevector: FourVector,
     pol: Polarization,
+    bandwidth: f64,
 }
 
 impl FocusedLaser {
@@ -24,8 +27,16 @@ impl FocusedLaser {
             waist,
             duration,
             wavevector,
-            pol
+            pol,
+            bandwidth: 0.0,
         }
+    }
+
+    pub fn with_finite_bandwidth(self) -> Self {
+        let mut cpy = self;
+        let n_fwhm = SPEED_OF_LIGHT * cpy.duration * cpy.wavevector[0] / (2.0 * consts::PI);
+        cpy.bandwidth = (0.5 * consts::LN_2).sqrt() / (consts::PI * n_fwhm);
+        cpy
     }
 
     fn omega(&self) -> f64 {
@@ -171,14 +182,28 @@ impl Field for FocusedLaser {
         (r_new, u_new)
     }
 
-    fn radiate<R: Rng>(&self, _r: FourVector, u: FourVector, dt: f64, rng: &mut R) -> Option<FourVector> {
-        let kappa = SPEED_OF_LIGHT * COMPTON_TIME * self.wavevector;
+    fn radiate<R: Rng>(&self, _r: FourVector, u: FourVector, dt: f64, rng: &mut R) -> Option<(FourVector, FourVector)> {
+        let width = 1.0 + self.bandwidth * rng.sample::<f64,_>(StandardNormal);
+        assert!(width > 0.0, "The fractional bandwidth of the pulse, {:.3e}, is large enough that the sampled frequency has fallen below zero!", self.bandwidth);
+        let kappa = SPEED_OF_LIGHT * COMPTON_TIME * self.wavevector * width;
         let prob = nonlinear_compton::probability(kappa, u, dt).unwrap_or(0.0);
         if rng.gen::<f64>() < prob {
-            let (_n, k) = nonlinear_compton::generate(kappa, u, rng, None);
-            Some(k)
+            let (n, k) = nonlinear_compton::generate(kappa, u, rng, None);
+            Some((k, u + (n as f64) * kappa - k))
         } else {
             None
+        }
+    }
+
+    fn pair_create<R: Rng>(&self, r: FourVector, ell: FourVector, dt: f64, rng: &mut R, rate_increase: f64) -> (f64, Option<(FourVector, FourVector)>) {
+        let a = self.a_sqd(r).sqrt();
+        let kappa = SPEED_OF_LIGHT * COMPTON_TIME * self.wavevector;
+        let prob = pair_creation::probability(ell, kappa, a, dt).unwrap_or(0.0);
+        if rng.gen::<f64>() < prob * rate_increase {
+            let (n, q_p) = pair_creation::generate(ell, kappa, a, rng);
+            (prob, Some((ell + (n as f64) * kappa - q_p, q_p)))
+        } else {
+            (prob, None)
         }
     }
 }