+++ docs/changelog.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## v0.7.0
+
+In progress.
+
+New:
+
+* Finite bandwidth correction: optionally account for the effect of the laser pulse duration when sampling photon momenta.
+* Electron-positron pair creation, via LMA and LCFA rates. `leading-order-only` feature, if enabled, replaces the LMA rate with a perturbative equivalent.
+* Previous behaviour, i.e. no pair creation or photon tracking, is recovered by running with `pair_creation: false` in the input file.
+* Photon primaries, with either Gaussian or bremsstrahlung energy spectra.
+* Automatic naming of output files.
+
+Fixed:
+
+* Build failure caused by missing git info.
+
+Removed:
+
+* `no-radiation-reaction` feature. Recoil on photon emission can now be disabled via the input file.
+
 ## v0.6.2
 
 Fixed: